chore: update the release instruction
